refactor(ui): organize components into subfolders closes #563
Move all root-level components into themed subdirectories:
- common: attribution-badges, spinner, stat-card, tabs
- dialogs: delete-dialog, title-lookup-results-dialog
- forms: book/game/movie/music forms
- images: image-display, image-lookup-info, image-upload, image-url-preview
- inputs: multiselect-input, singleselect-input, sortable-column-header
- layout: error-screen, footer, pending-navigation
- scan: barcode-scanner, scan-button

Update all import paths across routes and components.

[AI-assisted]
