+++ MediaSet.Remix/app/components/images/image-display.test.tsx
