+++ MediaSet.Remix/app/components/images/image-display.tsx
