+++ MediaSet.Remix/app/components/scan/barcode-scanner.test.tsx
