+++ MediaSet.Remix/app/components/scan/barcode-scanner.tsx
