+++ MediaSet.Remix/app/routes/$entity.add/route.tsx
@@ -18,11 +18,11 @@ import { formToDto, getEntityFromParams, singular } from '~/utils/helpers';
 import { BookEntity, Entity, GameEntity, MusicEntity, MovieEntity } from '~/models';
 import { getLookupCapabilities, isBarcodeLookupAvailable } from '~/api/lookup-capabilities-data';
 import { serverLogger } from '~/utils/serverLogger';
-import BookForm, { BookLookupSection } from '~/components/book-form';
-import MovieForm, { MovieLookupSection } from '~/components/movie-form';
-import GameForm, { GameLookupSection } from '~/components/game-form';
-import MusicForm, { MusicLookupSection } from '~/components/music-form';
-import TitleLookupResultsDialog from '~/components/title-lookup-results-dialog';
+import BookForm, { BookLookupSection } from '~/components/forms/book-form';
+import MovieForm, { MovieLookupSection } from '~/components/forms/movie-form';
+import GameForm, { GameLookupSection } from '~/components/forms/game-form';
+import MusicForm, { MusicLookupSection } from '~/components/forms/music-form';
+import TitleLookupResultsDialog from '~/components/dialogs/title-lookup-results-dialog';
 // Server-only lookup utilities are imported dynamically inside the action to avoid client bundling
 
 function isLookupError(result: unknown): result is { message: string; statusCode: number } {
